+++ test/release_updater_test.dart
@@ -1,3 +1,4 @@
+@TestOn('vm')
 import 'dart:async';
 import 'dart:convert' as dart_convert;
 import 'dart:io';